// src/app/shared/model/userAccessDTO.ts
/**
 * CovidCare API
 * CovidCare API documentation
 *
 * OpenAPI spec version: 0.0.1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
 import { SecurityUserDTO } from './securityUserDTO';

export interface UserAccessDTO {
    accessId?: number;
    id?: number;
    lastModified?: Date;
    lastModifiedBy?: string;
    level?: UserAccessDTO.LevelEnum;
    securityUserId?: number;
    securityUserLogin?: String;
    securityUser:SecurityUserDTO;
}
export namespace UserAccessDTO {
    export type LevelEnum = 'HOSPITAL' | 'SUPPLIER' | 'TALUKA' | 'DISTRICT' | 'STATE' | 'MUNCIPALCORP';
    export const LevelEnum = {
        HOSPITAL: 'HOSPITAL' as LevelEnum,
        SUPPLIER: 'SUPPLIER' as LevelEnum,
        TALUKA: 'TALUKA' as LevelEnum,
        DISTRICT: 'DISTRICT' as LevelEnum,
        STATE: 'STATE' as LevelEnum,
        MUNCIPALCORP: 'MUNCIPALCORP' as LevelEnum
    };
}